Extract metadata and line ranges
